Got GameBoard rendering and displaying
Fixed the issue where the GameBoard was not rendering by taking the
definition of the boardJsx out of the useEffect function.
Created a helper function for replacing any empty values in a board
array with empty strings because .map doesn't work on empty values in
an array.
Setup cells to be input elements with a type of number, a min value of
1 and a max value of 9 to reflect that data that will be entered.
Styled the board as an HTML grid with 9 columns.