Implemented GameShow components and Delete action
Implemented the GameShow component for showing a single Game and updated
GameIndex to call the GameShow component.
Added a Delete button to the GameShow component.
Created a function which sends a request to delete the specified game to
the API and updates the state of the GameIndex component.
Updated styling to remove buttons from the input cells.