Updated gameboard formatting and input method
Created a Cell component for each of the cells.
Added a modal to the Cell component that appears when the user clicks on
a cell. The modal contains 9 buttons labeled 1-9 that the user can pick
from to update the selected cell.
Updated the handleChange method to account for the new input method.
Updated the styling of the board to make it bigger and easier to read.
Updated the README for completed tasks and new tasks.